backend-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
redis==5.2.1
psycopg2-binary==2.9.10
